Week 6
This week we worked with functions. Functions in javascript are similar to the ones in java, which I know very well.

We looked at what happens when you add too few arguments to a function call. We also looked at how to call anonymous functions and the difference between them and normal functions.